Split target description schema to its own crate.
