+++ probe-rs/src/config/mod.rs
@@ -22,26 +22,21 @@
 //! be used to read targets from a YAML file.
 //!
 
-mod chip;
-mod chip_family;
 mod chip_info;
 mod flash_algorithm;
-mod flash_properties;
-mod memory;
 mod registry;
 mod target;
 
-pub use chip::Chip;
-pub use chip_family::ChipFamily;
-pub use flash_algorithm::{FlashAlgorithm, RawFlashAlgorithm};
-pub use flash_properties::FlashProperties;
-pub use memory::{
-    MemoryRange, MemoryRegion, NvmRegion, PageInfo, RamRegion, SectorDescription, SectorInfo,
+pub use probe_rs_target::{
+    Chip, ChipFamily, FlashProperties, MemoryRange, MemoryRegion, NvmRegion, PageInfo, RamRegion,
+    RawFlashAlgorithm, SectorDescription, SectorInfo, TargetDescriptionSource,
 };
+
+pub use flash_algorithm::FlashAlgorithm;
 pub use registry::{
     add_target_from_yaml, families, get_target_by_name, search_chips, RegistryError,
 };
-pub use target::{Target, TargetDescriptionSource, TargetParseError, TargetSelector};
+pub use target::{Target, TargetParseError, TargetSelector};
 
 // Crate-internal API
 pub(crate) use chip_info::ChipInfo;
